requirements.txt
aiohttp
aiohttp-apispec
pyyaml